core/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
cssselect2==0.9.0
debugpy==1.8.20
decorator==5.2.1
dj-database-url==3.1.2
Django==5.2.11
django-allauth==65.15.0
django-crispy-forms==2.6
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.20.0
fonttools==4.62.0
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.12.0
gunicorn==25.1.0
html5lib==1.1
idna==3.11
implicit==0.7.2
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
oscrypto==1.3.0
packaging==26.0
pandas==2.3.3
parso==0.8.6
pillow==12.1.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pycairo==1.29.0
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
pyHanko==0.34.1
pyhanko-certvalidator==0.30.1
PyJWT==2.11.0
pypdf==6.8.0
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
razorpay==2.0.1
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
svglib==1.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.1.0
torch==2.10.0+cpu
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
twilio==9.10.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.12.0
xhtml2pdf==0.2.17
yarl==1.22.0
zopfli==0.4.1
